+++ eqcorrscan/core/lag_calc.py
@@ -449,6 +449,7 @@ def _prepare_data(family, detect_data, shift_len, all_vert=False,
         Logger.info("Setting length to {0}s to give an integer number of "
                     "samples".format(length))
     prepick = shift_len + family.template.prepick
+    Logger.debug(f"prepick: {prepick} using shift_len={shift_len}, template.prepick={family.template.prepick}")
     detect_streams_dict = family.extract_streams(
         stream=detect_data, length=length, prepick=prepick,
         all_vert=all_vert, all_horiz=all_horiz, vertical_chans=vertical_chans,