+++ eqcorrscan/utils/findpeaks.py
@@ -10,11 +10,17 @@
 import ctypes
 import random
 import logging
+import warnings
+
 import numpy as np
+
+from typing import List, Union
 
 from multiprocessing import Pool, cpu_count
 from concurrent.futures import ThreadPoolExecutor
 from scipy import ndimage
+from obspy import UTCDateTime
+from obspy.core.event import Catalog, Event
 
 from eqcorrscan.utils.correlate import pool_boy
 from eqcorrscan.utils.libnames import _load_cdll
@@ -478,6 +484,231 @@ def coin_trig(peaks, stachans, samp_rate, moveout, min_trig, trig_int):
         return output
     else:
         return []
+
+
+# ##################### Declustering based on pick-time #######################
+
+
+def _pick_time(
+    event: Event,
+    station: str,
+    phase_hint: str,
+    zero_time: UTCDateTime
+) -> float:
+    """
+    Get the average pick time for a station for an event,
+    relative to a reference time.
+
+    :param event: Event to read picks from
+    :station: Station to get picks for
+    :phase_hint: Phase hint to get picks for
+    :zero_time: Reference time to compare to
+
+    :returns: Pick time relative to zero-time in seconds.
+    """
+    picks = [p for p in event.picks
+             if p.waveform_id.station_code == station
+             and p.phase_hint == phase_hint]
+    if len(picks) == 0:
+        return np.nan
+    pick_times = [p.time - zero_time for p in picks]
+    if len(pick_times) == 1:
+        return pick_times[0]
+    return sum(pick_times) / len(pick_times)
+
+
+def average_pick_time_diff_matrix(
+    catalog: Union[Catalog, List[Event]],
+    compiled: bool = True
+) -> np.ndarray:
+    """
+    Compute the average pick time difference matrix for events vs events
+
+    :param catalog: Catalog of events to compare pick times for
+    :compiled: Whether to use internal compiled code (faster) or not.
+
+    :returns: Pick time difference matrix - square.
+    """
+    # Make arrays of pick times for each phase and each station
+    sta_phases = list(
+        {f"{p.waveform_id.station_code}_{p.phase_hint}"
+         for ev in catalog for p in ev.picks if p.phase_hint in "PS"})
+    # Using a list for sta_phases to retain order
+    pick_arrays = dict()
+    zero_time = (catalog[0].preferred_origin() or catalog[0].origins[-1]).time
+    # This bit is not very fast, but gets the arrays in C shape
+    for sta_ph in sta_phases:
+        pick_arrays.update({
+            sta_ph: np.array(
+                [_pick_time(ev, *sta_ph.split('_'), zero_time=zero_time)
+                 for ev in catalog])})
+    if compiled:
+        return _compute_matrix_c(
+            pick_arrays=pick_arrays, n_events=len(catalog))
+    else:
+        return _compute_matrix(
+            pick_arrays=pick_arrays, n_events=len(catalog))
+
+
+def _compute_matrix(pick_arrays: dict, n_events: int) -> np.ndarray:
+    out = np.zeros((n_events, n_events))
+    for i in range(n_events):
+        # Matrix is symmetric, so we can just do upper or lower half.
+        for j in range(i + 1, n_events):
+            # We get occasional RuntimeWarnings about mean of empty slice
+            # when no stations are shared. nan is returned for this case.
+            with warnings.catch_warnings():
+                warnings.simplefilter("ignore")
+                pick_delta = np.nanmean(
+                    [pick_arrays[sta_ph][i] - pick_arrays[sta_ph][j]
+                     for sta_ph in pick_arrays.keys()])
+            out[i, j] = pick_delta
+    out -= out.T
+    return np.nan_to_num(out, nan=np.inf)
+
+
+def _compute_matrix_c(pick_arrays: dict, n_events: int) -> np.ndarray:
+    out = np.ascontiguousarray(
+        np.zeros(n_events * n_events), np.float64)
+    pick_array = np.ascontiguousarray(
+        np.concatenate([value for value in pick_arrays.values()]), np.float64)
+
+    utilslib = _load_cdll("libutils")
+    utilslib.average_pick_time_diff_matrix.argtypes = [
+        ctypes.c_int, ctypes.c_int,
+        np.ctypeslib.ndpointer(dtype=np.float64, shape=(len(pick_array), ),
+                               flags='C_CONTIGUOUS'),
+        np.ctypeslib.ndpointer(dtype=np.float64, shape=(len(out), ),
+                               flags='C_CONTIGUOUS')]
+    utilslib.average_pick_time_diff_matrix.restype = ctypes.c_int
+
+    ret = utilslib.average_pick_time_diff_matrix(
+        ctypes.c_int(len(pick_arrays.keys())), ctypes.c_int(n_events),
+        pick_array, out)
+
+    if ret != 0:
+        raise Exception("Error in internal C code.")
+
+    out = out.reshape(n_events, n_events)
+    out -= out.T
+    return np.nan_to_num(out, nan=np.inf)
+
+
+def get_det_val(event: Event) -> Union[float, None]:
+    """
+    Get the detection value for an event. Prefers sum of pick correlations.
+
+    :param event: Event to get detection value for.
+
+    :returns: Detection value.
+    """
+    det_val = get_comment_val(value_name="detect_val", event=event)
+    pick_sum = 0
+    for pick in event.picks:
+        pick_val = get_comment_val("cc_max", event=pick)
+        if pick_val:
+            pick_sum += pick_val
+    if det_val:
+        if pick_sum > det_val:
+            return pick_sum
+        else:
+            return det_val
+    elif pick_sum:
+        return pick_sum
+    return None
+
+
+def get_comment_val(value_name: str, event: Event) -> Union[float, None]:
+    """
+    Get value from EQcorrscan-style comment.
+
+    :param value_name: Comment string before value
+    :param event: Event to search comments for
+
+    :returns: Comment value.
+    """
+    value = None
+    for comment in event.comments:
+        if value_name in comment.text:
+            if "=" in comment.text:
+                # Should be a number
+                try:
+                    value = float(comment.text.split('=')[-1])
+                except ValueError:
+                    # Leave as a string
+                    break
+                except Exception as e:
+                    Logger.exception(
+                        f"Could not get {value_name} {comment.text} "
+                        f"due to {e}")
+                else:
+                    break
+            elif ":" in comment.text:
+                # Should be a string
+                try:
+                    value = comment.text.split(": ")[-1]
+                except Exception as e:
+                    Logger.exception(
+                        f"Could not get {value_name} from {comment.text} "
+                        f"due to {e}")
+                else:
+                    break
+    return value
+
+
+def decluster_pick_times(
+    catalog: Union[Catalog, List[Event]],
+    trig_int: float
+) -> Catalog:
+    """
+    Decluster a catalog based on average pick-time similarity.
+
+    If two events have (on average) pick-times within trig_int seconds of
+    each other, then the event with the highest detection value, or number
+    of picks (if detection values are not found) will be retained and the other
+    event discarded.
+
+    Useful for after lag-calc to remove events that have different template
+    locations, but stem from similar pick-times.
+
+    :param catalog: Catalog of events to decluster
+    :trig_int: Minimum inter-event time in seconds
+
+    :returns: Declustered catalog.
+    """
+    detect_vals = np.array([abs(get_det_val(ev) or len(ev.picks))
+                            for ev in catalog])
+    # Sort catalog in order of detect_vals - from largest to smallest
+    order = np.argsort(detect_vals)[::-1]
+    # detect_vals = detect_vals[order]
+    if isinstance(catalog, Catalog):
+        events = catalog.events
+    else:
+        events = catalog
+
+    events = [events[ind] for ind in order]
+
+    # This is obviously very slow at the moment
+    pick_matrix = average_pick_time_diff_matrix(catalog=events)
+    # We only care about the absolute value (+/- trig_int)
+    pick_matrix = np.abs(pick_matrix)
+
+    drop_events = np.zeros(len(events), dtype=bool)
+    # Loop through from highest detection value to lowest - this way we keep
+    # the best detections
+    for i, ev in enumerate(events):
+        # If we are dropping this event then carry on
+        if drop_events[i]:
+            continue
+        # Find any events within trig int and set their drop status to True
+        drop_mask = pick_matrix[i] <= trig_int
+        # Make sure we retain the "core" event in this loop
+        drop_mask[i] = False
+        drop_events[drop_mask] = True
+        # _dropped = [ev for j, ev in _dropped(events) if drop_mask[j]]
+
+    # Get the events that we want to keep!
+    return Catalog([ev for i, ev in enumerate(events) if not drop_events[i]])
 
 
 if __name__ == "__main__":