+++ CHANGES.md
@@ -4,6 +4,12 @@
     (`min_stations` kwarg)
 - core.match_filter.party
   - Add client_lag_calc method to run lag-calc using data from a client.
+- core.match_filter.detection
+  - Set detect-time correctly when earliest channel is unused in correlation
+- core.lag_calc
+  - More stable extraction of data for correlation
+- utils.findpeaks
+  - Added pick-time declustering.
 
 ## 0.5.0
 * core.match_filter.tribe